pyproject.toml
[project]
name = "customer_agent"
version = "0.1.0"
description = "An intelligent customer support agent using LangGraph"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "langgraph>=0.0.1",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.1",
    "python-dotenv>=1.0.0",
    "termcolor>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 